[adapters] Implement service adapters
