+++ adapters/cosmos-hub/cosmos-hub/index.ts
@@ -1,0 +1,10 @@
+import { ServiceInfo } from '@/_lib/types';
+
+const info: ServiceInfo = {
+  native: true,
+  name: 'Cosmos Hub',
+  logo: '',
+  url: 'https://cosmos.network',
+  description: 'The First of Thousands of Interconnected Blockchains',
+};
+export default info;
