+++ adapters/klaytn/klaystation/index.ts
@@ -1,7 +1,13 @@
-import { PROTOCOL_ADDRESS } from './constants';
-import { getDelegations } from './staking';
+import { ServiceInfo } from '@/_lib/types';
 
-export const KlayStation = {
-  PROTOCOL_ADDRESS,
-  getDelegations,
+const info: ServiceInfo = {
+  native: false,
+  name: 'KLAYSTATION',
+  logo: '/assets/icons/services/klaytn/klaystation.svg',
+  url: 'https://klaystation.io/',
+  description: {
+    en: 'KLAYSTATION is a staking tool that allows users to delegate KLAY to GCs(Validators) and get rewards.',
+    ko: '클레이스테이션은 GC 노드(밸리데이터)들에게 KLAY를 위임하고 보상을 수령할 수 있도록 하는 스테이킹 툴입니다.',
+  },
 };
+export default info;
