+++ adapters/klaytn/klayswap/index.ts
@@ -1,14 +1,13 @@
-import { KSP_ADDRESS, KSP_TOKEN_INFO, VOTING_KSP_ADDRESS } from './constants';
-import { getGovernanceStake } from './governance';
-import { getLPPoolBalance } from './lp';
-import { getLeveragePoolList, getSinglePoolBalance } from './single';
+import { ServiceInfo } from '@/_lib/types';
 
-export const KlaySwap = {
-  KSP_ADDRESS,
-  KSP_TOKEN_INFO,
-  VOTING_KSP_ADDRESS,
-  getGovernanceStake,
-  getLPPoolBalance,
-  getSinglePoolBalance,
-  getLeveragePoolList,
+const info: ServiceInfo = {
+  native: false,
+  name: 'KLAYswap',
+  logo: '',
+  url: 'https://klayswap.com',
+  description: {
+    en: 'KLAYswap is an AMM-based Instant Swap Protocol.',
+    ko: '클레이스왑은 클레이튼 생태계에서 가장 커다란 DEX(AMM)입니다.',
+  },
 };
+export default info;
