+++ adapters/klaytn/kokonutswap/index.ts
@@ -1,19 +1,13 @@
-import {
-  KOKOS_ADDRESS,
-  KOKOS_TOKEN_INFO,
-  KSD_ADDRESS,
-  KSD_TOKEN_INFO,
-  STAKED_KOKOS_ADDRESS,
-} from './constants';
-import { getGovernanceStake } from './governance';
-import { getLPPoolBalance } from './lp';
+import { ServiceInfo } from '@/_lib/types';
 
-export const KokonutSwap = {
-  KSD_ADDRESS,
-  KSD_TOKEN_INFO,
-  KOKOS_ADDRESS,
-  KOKOS_TOKEN_INFO,
-  STAKED_KOKOS_ADDRESS,
-  getLPPoolBalance,
-  getGovernanceStake,
+const info: ServiceInfo = {
+  native: false,
+  name: 'Kokonut Swap',
+  logo: '',
+  url: 'https://kokonutswap.finance',
+  description: {
+    en: 'Kokonut Swap is a next-generation AMM DEX protocol for the Klaytn ecosystem, which enables users/protocols to exchange stablecoins with low slippage.',
+    ko: '코코넛 스왑은 유저 및 프로토콜들이 스테이블코인을 낮은 슬리피지로 교환할 수 있게 해주는 클레이튼 생태계의 2세대 AMM DEX 프로토콜입니다.',
+  },
 };
+export default info;
