+++ adapters/osmosis/osmosis/index.ts
@@ -1,0 +1,13 @@
+import { ServiceInfo } from '@/_lib/types';
+
+const info: ServiceInfo = {
+  native: true,
+  name: 'Osmosis',
+  logo: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/osmosis/images/osmo.png',
+  url: 'https://osmosis.zone',
+  description: {
+    en: 'Osmosis is the largest interchain DEX in the Cosmos Ecosystem.',
+    ko: '오스모시스는 코스모스 생태계 최초, 최대의 인터체인 DEX입니다.',
+  },
+};
+export default info;
